gateway_service/requirements.txt
fastapi
uvicorn
httpx
pydantic
python-jose[cryptography]